requirements.txt
python-dotenv
tencentcloud-sdk-python
